move base class back to __init__
