+++ capreolus/collection/antique.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 
-from .base import Collection
+from . import Collection
 from capreolus import ModuleBase, Dependency, ConfigOption, constants
 from capreolus.utils.common import download_file, hash_file
 from capreolus.utils.loginit import get_logger
