+++ capreolus/collection/codesearchnet.py
@@ -3,7 +3,7 @@ import pickle
 from tqdm import tqdm
 from zipfile import ZipFile
 
-from .base import Collection
+from . import Collection
 from capreolus import ModuleBase, Dependency, ConfigOption, constants
 from capreolus.utils.common import download_file, hash_file, remove_newline
 from capreolus.utils.loginit import get_logger
