+++ capreolus/collection/covid.py
@@ -7,7 +7,7 @@ from tqdm import tqdm
 from pathlib import Path
 import pandas as pd
 
-from .base import Collection
+from . import Collection
 from capreolus import ModuleBase, Dependency, ConfigOption, constants
 from capreolus.utils.common import download_file, hash_file, remove_newline
 from capreolus.utils.loginit import get_logger
