+++ capreolus/extractor/embedtext.py
@@ -7,7 +7,7 @@ import tensorflow as tf
 from pymagnitude import Magnitude, MagnitudeUtils
 from tqdm import tqdm
 
-from .base import Extractor
+from . import Extractor
 from capreolus import ModuleBase, Dependency, ConfigOption, constants, get_logger
 from capreolus.utils.common import padlist
 from capreolus.utils.exceptions import MissingDocError
