+++ capreolus/reranker/DSSM.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
-from profane import ConfigOption, Dependency
 
+from capreolus import ConfigOption, Dependency
 from capreolus.extractor.bagofwords import BagOfWords
 from capreolus.reranker import Reranker
 from capreolus.utils.loginit import get_logger
