+++ capreolus/reranker/DUET.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
-from profane import ConfigOption, Dependency
 
+from capreolus import ConfigOption, Dependency
 from capreolus.reranker import Reranker
 from capreolus.reranker.common import create_emb_layer
 from capreolus.utils.loginit import get_logger
