+++ capreolus/reranker/TFKNRM.py
@@ -1,6 +1,6 @@
 import tensorflow as tf
-from profane import ConfigOption, Dependency
 
+from capreolus import ConfigOption, Dependency
 from capreolus.reranker import Reranker
 from capreolus.reranker.common import RbfKernelBankTF, similarity_matrix_tf
