+++ capreolus/reranker/TFVanillaBert.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
-from profane import ConfigOption, Dependency
 from transformers import TFBertForSequenceClassification
 
+from capreolus import ConfigOption, Dependency
 from capreolus.reranker import Reranker
 from capreolus.utils.loginit import get_logger
