--- capreolus/tokenizer/base.py
@@ -1,5 +1,0 @@
-from capreolus import ModuleBase, Dependency, ConfigOption
-
-
-class Tokenizer(ModuleBase):
-    module_type = "tokenizer"
